+++ src/gui/input/SpaceNavInputDriver.h
@@ -43,10 +43,10 @@ public:
   const std::string& get_name() const override;
   std::string get_info() const override;
 
-  int getButtonCount() const override {
+  size_t getButtonCount() const override {
     return InputDriver::max_buttons;
   }
-  int getAxisCount() const override {
+  size_t getAxisCount() const override {
     return InputDriver::max_axis;
   }
